faster turbo, move token utils to its own file (#277)
