+++ connect/src/routes/resolver.ts
@@ -1,28 +1,11 @@
-import { Chain, Network } from "@wormhole-foundation/sdk-base";
-import {
-  ChainContext,
-  TokenId,
-  canonicalAddress,
-  resolveWrappedToken,
-} from "@wormhole-foundation/sdk-definitions";
+import { Network } from "@wormhole-foundation/sdk-base";
+import { ChainContext, TokenId, resolveWrappedToken } from "@wormhole-foundation/sdk-definitions";
 import { Wormhole } from "../wormhole";
 import { RouteTransferRequest } from "./request";
 import { Route, RouteConstructor, isAutomatic } from "./route";
+import { uniqueTokens } from "./token";
 
 export type RouteSortOptions = "cost" | "speed";
-
-export function uniqueTokens<C extends Chain>(tokens: TokenId<C>[]): TokenId<C>[] {
-  if (tokens.length === 0) return [];
-
-  // take the first chain, all should be equal
-  const { chain } = tokens[0]!;
-
-  if (!tokens.every((t) => t.chain === chain)) throw new Error("Not every chain is equal");
-
-  return Array.from(new Set(tokens.map((t) => canonicalAddress(t)))).map((a) =>
-    Wormhole.tokenId(chain, a),
-  );
-}
 
 export class RouteResolver<N extends Network> {
   wh: Wormhole<N>;
@@ -37,9 +20,8 @@ export class RouteResolver<N extends Network> {
   async supportedSourceTokens(chain: ChainContext<Network>): Promise<TokenId[]> {
     if (this.inputTokenList) return this.inputTokenList;
     const itl = await Promise.all(
-      this.routeConstructors.flatMap(async (rc) => rc.supportedSourceTokens(chain)),
+      this.routeConstructors.map(async (rc) => rc.supportedSourceTokens(chain)),
     );
-
     this.inputTokenList = uniqueTokens(itl.flat());
     return this.inputTokenList!;
   }
