+++ connect/src/routes/token.ts
@@ -1,0 +1,54 @@
+import { Chain, Network, tokens } from "@wormhole-foundation/sdk-base";
+import {
+  ChainContext,
+  TokenId,
+  canonicalAddress,
+  isNative,
+  isTokenId,
+} from "@wormhole-foundation/sdk-definitions";
+import { Wormhole } from "../wormhole";
+
+export interface TokenDetails {
+  id: TokenId;
+  decimals: number;
+  symbol?: tokens.TokenSymbol;
+  wrapped?: TokenId;
+}
+
+export function uniqueTokens<C extends Chain>(tokens: TokenId<C>[]): TokenId<C>[] {
+  if (tokens.length === 0) return [];
+
+  // take the first chain, all should be equal
+  const { chain } = tokens[0]!;
+  if (!tokens.every((t) => t.chain === chain)) throw new Error("Not every chain is equal");
+
+  return Array.from(new Set(tokens.map((t) => canonicalAddress(t)))).map((a) =>
+    Wormhole.tokenId(chain, a),
+  );
+}
+
+export function tokenAddresses(tokens: TokenId[]): string[] {
+  return tokens.map((t) => canonicalAddress(t));
+}
+
+export async function getTokenDetails<N extends Network>(
+  chain: ChainContext<N>,
+  token: TokenId,
+): Promise<TokenDetails> {
+  const address = isTokenId(token) ? canonicalAddress(token) : token;
+
+  const details = chain.config.tokenMap
+    ? tokens.filters.byAddress(chain.config.tokenMap!, address)
+    : undefined;
+
+  const symbol = details ? details.symbol : undefined;
+  const wrapped = isNative(token.address) ? await chain.getNativeWrappedTokenId() : undefined;
+  const decimals = Number(await chain.getDecimals(isTokenId(token) ? token.address : token));
+
+  return {
+    id: token,
+    decimals,
+    wrapped,
+    symbol,
+  };
+}
