+++ package.json
@@ -25,8 +25,8 @@
   },
   "scripts": {
     "lint": "npx turbo lint",
-    "build": "npx turbo build",
-    "rebuild": "npx turbo build --force",
+    "build": "npx turbo build --concurrency 20",
+    "rebuild": "npx turbo build --no-cache --force --concurrency 20",
     "test": "NETWORK=Mainnet npx turbo test",
     "prettier": "npm run prettier --workspaces --if-present",
     "docs": "NODE_OPTIONS=--max_old_space_size=8192 typedoc --options typedoc.json",
